move create asteroid to function in separate file
