+++ lib/index.js
@@ -2,23 +2,15 @@ const $ = require('jquery')
 const Ship = require('../lib/ship');
 const Asteroid = require('../lib/asteroid');
 const collisionDetection = require('../lib/collisions');
+const asteroids = require('../lib/asteroids');
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
 var ship = new Ship();
 
-function getRandomNumber(min, max){
-  return Math.floor(Math.random() * (max - min + 1)) + min;
-};
-
-// create Asteroids
-var asteroids = [];
-for (var x=0; x<5; x++){
-  var newX = getRandomNumber(0, canvas.width)
-  var newY = getRandomNumber(0, canvas.height)
-  var newAsteroid = new Asteroid(newX, newY)
-  asteroids.push(newAsteroid)
-};
+// function getRandomNumber(min, max){
+//   return Math.floor(Math.random() * (max - min + 1)) + min;
+// };
 
 var img = new Image();
 img.src = 'assets/images/nebula.png';
